requirements.txt
Flask==0.10.1
Flask-Misaka==0.3.0
Flask-SQLAlchemy==2.1
misaka==1.0.2
Pillow==3.0.0
